requirements.txt
# Core packages
requests>=2.25.1
urllib3>=1.26.5
icrawler>=0.6.6

# Web automation
selenium>=4.0.0
webdriver-manager>=3.5.2
fake-useragent>=0.1.11

# Image processing
Pillow>=8.3.1
imagehash>=4.2.1

# Hugging Face integration
huggingface_hub>=0.10.0

# Environment variables
python-dotenv>=0.19.0

# SSL & Security
pyOpenSSL>=20.0.1
cryptography>=3.4.8

# Utility
tqdm>=4.62.3